refactor(server): move position and subdomain to own modules
